refactor: Update component imports in page.tsx and Admin-Dashboard.tsx
